Added prompts

+++ 1-install.sh
@@ -1,24 +1,39 @@
 #!/bin/bash
 
-echo 
-echo "Installing packages from applist.txt file..."
-sleep 1
 
-xargs sudo apt install -y < applist.txt
+# Installing packages
+read -p "Do you wish to install packages from applist.txt? (y/n): " answer
 
-echo 
-echo "Installing fonts..."
-sleep 1
-
-stow fonts
-
-echo 
-echo "Changing default shell to Zsh..."
-sleep 1
-chsh -s $(which zsh)
-
+if [ "$answer" = "y" ]; then
+    echo "Installing packages from applist.txt file..."
+    xargs sudo apt install -y < applist.txt
+	echo "Success!"
+fi
 echo
+
+
+# Installing fonts
+read -p "Do you wish to install fonts? (y/n): " answer
+
+if [ "$answer" = "y" ]; then
+	echo "Installing fonts..."
+	stow fonts
+	echo "Success!"
+fi
+echo
+
+
+# Change default shell
+read -p "Do you wish to change the default shell to Zsh? (y/n): " answer
+
+if [ "$answer" = "y" ]; then
+	echo "Changing default shell to Zsh..."
+	chsh -s $(which zsh)
+	echo "Success!"
+fi
+echo
+
+# EXTRAS
 echo "NOTE:"
 echo "- Change your terminal's font to MesloLGS, size 14-16."
 echo "- Please log off to refresh the default shell!"
-sleep 2
